+++ src/lib.rs
@@ -6,13 +6,14 @@ mod compute_reduce1;
 mod compute_reduce2;
 mod compute_with_reduction;
 mod input_builder;
-mod input_builder0;
-mod input_builder1;
-mod input_builder2;
+mod input_builder0_zzz;
+mod input_builder1_zzz;
+mod input_builder2_zzz;
+mod input_builder_zzz;
 mod reduce;
 mod type_sequence;
 
 pub use composable::Composable;
 pub use compute::Compute;
-pub use input_builder::InputBuilder;
+pub use input_builder_zzz::InputBuilderZzz;
 pub use reduce::Reduce;
